+++ client/routes.js
@@ -2,11 +2,16 @@ import React, {Component} from 'react'
 import {connect} from 'react-redux'
 import {withRouter, Route, Switch} from 'react-router-dom'
 import PropTypes from 'prop-types'
-import { Login, Signup, UserHome, UserAccount } from './components'
+import {
+  Login,
+  Signup,
+  UserHome,
+  UserAccount,
+  AllProducts,
+  SingleProduct,
+  AddProduct
+} from './components'
 import {me} from './store'
-import AllProducts from './components/AllProducts'
-import SingleProduct from './components/SingleProduct'
-import AddProduct from './components/AddProduct'
 
 /**
  * COMPONENT
